refactor(preview): replace monolithic file with focused modules
preview-browser.ts (2084 lines) -> preview/ directory:
- preview-session.ts: session state and types
- preview-lifecycle.ts: BrowserView create/hide/park/dispose
- preview-navigation.ts: sync and navigation IPC handlers
- preview-capture.ts: screenshot capture and guest context
- preview-overlay.ts: region drag and element pick overlays
- preview-spill.ts: capture spill file management
- index.ts: thin wiring entrypoint
